llm-tutor/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-tutor-backend"
version = "0.1.0"
description = "LLM Tutor Backend (FastAPI)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.5",
  "pydantic-settings>=2.2",
  "python-multipart>=0.0.9",
  "email-validator>=2.0",
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.1",
  "asyncpg>=0.29",
  "pgvector>=0.3.3",
  "alembic>=1.13",
  "passlib[bcrypt]>=1.7",
  "pyjwt[crypto]>=2.8",
  "redis>=5.0",
  "celery[redis]>=5.3",
  "httpx>=0.27",
  "orjson>=3.10",
  "sse-starlette>=2.1",
  "itsdangerous>=2.1",
  "opentelemetry-sdk>=1.24",
  "opentelemetry-instrumentation-fastapi>=0.45b0",
  "opentelemetry-exporter-otlp>=1.24",
  "sentry-sdk>=2.5",
  "pypdf>=4.3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2", "pytest-asyncio>=0.23", "httpx>=0.27",
  "ruff>=0.5", "mypy>=1.11"
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
